Moving API classes to server, as they aren't used on the client.
