+++ PersistentMapAPI/PersistentMapAPI/PersistentMapAPI.csproj
@@ -49,8 +49,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Objects\FactionControl.cs" />
-    <Compile Include="Objects\FactionShop.cs" />
-    <Compile Include="Objects\HistoryResult.cs" />
     <Compile Include="Objects\MissionResult.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Objects\StarMap.cs" />
